+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- calculations on dates
+
 ## [5.1.4] - 2023-10-23
 
 ### Dependencies