+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `task`, `taskGroup` and `taskGroupInstance` props to `<OneBlinkForm />`
+
 ## [5.2.0] - 2023-10-25
 
 ### Added